:recycle: rename `fowg` binary to `flowg-server`
